+++ feat/tutorial/src/main/java/com/emotionstorage/tutorial/ui/component/PagerWithIndicator.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.tutorial.ui.tutorial
+package com.emotionstorage.tutorial.ui.component
 
 import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.animation.core.tween
@@ -23,7 +23,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
