+++ feat/tutorial/src/main/java/com/emotionstorage/tutorial/ui/onBoarding/ExpectationsScreen.kt
@@ -33,6 +33,7 @@ import com.emotionstorage.tutorial.R
 import com.emotionstorage.tutorial.presentation.onBoarding.ExpectationsEvent
 import com.emotionstorage.tutorial.presentation.onBoarding.ExpectationsViewModel
 import com.emotionstorage.tutorial.presentation.onBoarding.ExpectationsViewModel.State
+import com.emotionstorage.tutorial.ui.component.OnBoardingTitle
 import com.emotionstorage.ui.component.button.CtaButton
 import com.emotionstorage.ui.component.appBar.TopAppBar
 import com.emotionstorage.ui.theme.MooiTheme
