+++ feat/tutorial/src/main/java/com/emotionstorage/tutorial/ui/onBoarding/GenderBirthScreen.kt
@@ -29,6 +29,7 @@ import com.emotionstorage.tutorial.R
 import com.emotionstorage.tutorial.presentation.onBoarding.GenderBirthEvent
 import com.emotionstorage.tutorial.presentation.onBoarding.GenderBirthViewModel
 import com.emotionstorage.tutorial.presentation.onBoarding.GenderBirthViewModel.State
+import com.emotionstorage.tutorial.ui.component.OnBoardingTitle
 import com.emotionstorage.ui.component.button.CtaButton
 import com.emotionstorage.ui.component.picker.ScrollPicker
 import com.emotionstorage.ui.component.appBar.TopAppBar
